add arm64 asm

+++ sum_asm.go
@@ -1,6 +1,6 @@
-// +build amd64
+// +build amd64 arm64
 
 package checksum
 
 //go:noescape
-func sum(s uint, b []byte) uint
+func sum(s uint64, b []byte) uint64
